+++ packages/yourloops/test/integration/assert/team-dialog.assert.ts
@@ -29,14 +29,16 @@ import TeamAPI from '../../../lib/team/team.api'
 import { fireEvent, screen, waitFor, within } from '@testing-library/react'
 import { PhonePrefixCode } from '../../../lib/utils'
 import userEvent from '@testing-library/user-event'
-import { mockTeamApiForTeamCreation } from '../mock/team.api.mock'
+import { mockTeamApiForTeamCreation, NEW_TEAM_ID } from '../mock/team.api.mock'
 import { type Router } from '../models/router.model'
+import PatientApi from '../../../lib/patient/patient.api'
+import { loggedInUserId } from '../mock/auth0.hook.mock'
 
 export const checkCreateCareTeamDialog = async () => {
   jest.spyOn(TeamAPI, 'createTeam').mockResolvedValue(undefined)
   const teamMenu = screen.getByLabelText('Open team selection menu')
   await userEvent.click(teamMenu)
-  await userEvent.click(screen.getByText('Create a new care team'))
+  await userEvent.click(screen.getByText('Create a care team'))
   const dialogTeam = screen.getByRole('dialog')
   const createTeamButton = within(dialogTeam).getByRole('button', { name: 'Create team' })
   const cancelButton = within(dialogTeam).getByRole('button', { name: 'Cancel' })
@@ -122,7 +124,7 @@ export const checkTeamCreationFailure = async (): Promise<void> => {
 
   const teamMenu = screen.getByLabelText('Open team selection menu')
   await userEvent.click(teamMenu)
-  await userEvent.click(screen.getByText('Create a new care team'))
+  await userEvent.click(screen.getByText('Create a care team'))
 
   const dialog = screen.getByRole('dialog')
   const nameInput = within(dialog).getByRole('textbox', { name: 'Name' })
@@ -166,7 +168,7 @@ export const checkTeamCreationSuccess = async (router: Router): Promise<void> =>
 
   const teamMenu = screen.getByLabelText('Open team selection menu')
   await userEvent.click(teamMenu)
-  await userEvent.click(screen.getByText('Create a new care team'))
+  await userEvent.click(screen.getByText('Create a care team'))
 
   const dialog = screen.getByRole('dialog')
   const nameInput = within(dialog).getByRole('textbox', { name: 'Name' })
@@ -205,4 +207,5 @@ export const checkTeamCreationSuccess = async (router: Router): Promise<void> =>
   await waitFor(() => {
     expect(within(teamMenu).getByText(teamName)).toBeVisible()
   })
+  expect(PatientApi.getPatientsForHcp).toHaveBeenCalledWith(loggedInUserId, NEW_TEAM_ID)
 }