+++ packages/yourloops/test/integration/pages/daily/daily-view.common.spec.tsx
@@ -143,7 +143,7 @@ describe('Daily view for anyone', () => {
 
       // This checks that we tried to generate a pdf
       expect(downloadLinkElement.download).toEqual(`yourloops-report-${patient2AsTeamMember.userId}.pdf`)
-      expect(downloadLinkElement.href.length).toBeGreaterThan(17950)
+      expect(downloadLinkElement.href.length).toBeGreaterThan(17949)
       expect(downloadLinkElement.href.length).toBeLessThan(18040)
       expect(downloadLinkElement.click).toHaveBeenCalledTimes(1)